CB-6795,CB-6796: Fix infer_shape of TopK and Range op
Align output dtype with onnx TopK/Range op and consider float16
input dtype for TopK and ArmTopK op.

+++ AIPUBuilder/Parser/ops/onnx_ops/array_ops.py
@@ -736,7 +736,7 @@ class RangeOp(OpHasOneOutPort, ConstLikeOp, OnnxOp):
         if any(val is None for val in [self.start, self.limit, self.delta]):
             out_tensor = None
         else:
-            out_tensor = np.arange(self.start, self.limit, self.delta)
+            out_tensor = np.arange(self.start, self.limit, self.delta).astype(self.start.dtype)
         self.set_out_tensor(out_tensor)
 
     def __getattr__(self, item):
